requirements.txt
django
pillow
requests
djangorestframework